+++ lib/isolator/database_cleaner_support.rb
@@ -7,6 +7,7 @@ require "database_cleaner/active_record/transaction"
     def start
       super
       connection_id = connection_class.connection.object_id
+      Isolator.set_current_connection_id(connection_id, context: :database_cleaner)
       Isolator.set_connection_threshold(
         Isolator.transactions_threshold(connection_id) + 1,
         connection_id
@@ -15,6 +16,7 @@ require "database_cleaner/active_record/transaction"
 
     def clean
       connection_id = connection_class.connection.object_id
+      return unless connection_id == Isolator.current_connection_id(context: :database_cleaner)
       Isolator.set_connection_threshold(
         Isolator.transactions_threshold(connection_id) - 1,
         connection_id